Rewrite how timeouts are configured and calculated
It's easier to configure now, while it accounts for different kinds of
requests - especially important is it to distinguish between requests
that scale with the amount of bytes of objects, and with the amuont of
objects in the operation.
The user of this library now has to specify the assumed time each unit
of work (byte or object) takes.